# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ajloun Vision Trend Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /points/award:
    post:
      operationId: awardPoints
      tags: [points]
      summary: Award points to a user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AwardPointsInput"
      responses:
        "200":
          description: Points awarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointsSummary"
...
  /points/users/{userId}/transactions:
    get:
      operationId: listUserTransactions
      tags: [points]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PointsTransaction"

  # ─── BADGES ────────────────────────────────────────────
  /badges:
    get:
      operationId: listBadges
      tags: [points]
      responses:
        "200":
          description: Badges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Badge"
    post:
      operationId: createBadge
      tags: [points]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BadgeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Badge"
  /badges/{id}:
    delete:
      operationId: deleteBadge
      tags: [points]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /badges/award:
    post:
      operationId: awardBadge
      tags: [points]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AwardBadgeInput"
      responses:
        "200":
          description: Awarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /badges/user/{userId}:
    get:
      operationId: getUserBadges
      tags: [points]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User badges
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Badge"

  # ─── REWARDS / STORE ───────────────────────────────────
  /rewards:
    get:
      operationId: listRewards
      tags: [rewards]
      responses:
        "200":
          description: Rewards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reward"
    post:
      operationId: createReward
      tags: [rewards]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewardInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reward"
  /rewards/{id}:
    patch:
      operationId: updateReward
      tags: [rewards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RewardInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reward"
    delete:
      operationId: deleteReward
      tags: [rewards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /rewards/{id}/claim:
    post:
      operationId: claimReward
      tags: [rewards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
...
  /rewards/claims/{id}/status:
    patch:
      operationId: updateClaimStatus
      tags: [rewards]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RewardClaim"

  # ─── NOTIFICATIONS ─────────────────────────────────────
...
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    # ─── AUTH ─────────────────────────────
...
    PointsSummary:
      type: object
      required: [userId, volunteerPoints, trainingPoints, activityPoints, totalPoints]
      properties:
        userId:
          type: integer
        volunteerPoints:
          type: integer
        trainingPoints:
          type: integer
        activityPoints:
          type: integer
        totalPoints:
          type: integer
    PointsTransaction:
      type: object
      required: [id, userId, type, category, points, description, descriptionAr, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        category:
          type: string
        points:
          type: integer
        description:
          type: string
        descriptionAr:
          type: string
        createdAt:
          type: string
    AwardPointsInput:
      type: object
      required: [userId, points, category, descriptionAr]
      properties:
        userId:
          type: integer
        points:
          type: integer
        category:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
...
    Badge:
      type: object
      required: [id, name, nameAr, description, descriptionAr, icon, color, requiredPoints, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        icon:
          type: string
        color:
          type: string
        requiredPoints:
          type: integer
        createdAt:
          type: string
    BadgeInput:
      type: object
      required: [name, nameAr, requiredPoints]
      properties:
        name:
          type: string
        nameAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        icon:
          type: string
        color:
          type: string
        requiredPoints:
          type: integer
    AwardBadgeInput:
      type: object
      required: [userId, badgeId]
      properties:
        userId:
          type: integer
        badgeId:
          type: integer

    # ─── REWARDS ──────────────────────────
    Reward:
      type: object
      required: [id, title, titleAr, description, descriptionAr, pointsCost, stock, isActive, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        pointsCost:
          type: integer
        stock:
          type: integer
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    RewardInput:
      type: object
      required: [title, titleAr, pointsCost, stock]
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        pointsCost:
          type: integer
        stock:
          type: integer
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
    RewardClaim:
      type: object
      required: [id, userId, rewardId, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        rewardId:
          type: integer
        rewardTitle:
          type: string
        rewardTitleAr:
          type: string
        status:
          type: string
        createdAt:
          type: string
    ClaimStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string

    # ─── NOTIFICATIONS ────────────────────
...