# lib/api-zod/src/generated/types/awardPointsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Ajloun Vision Trend Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AwardPointsInput {
  userId: number;
  points: number;
  category: string;
  description?: string;
  descriptionAr: string;
}